Use the CSS Transform scale Property to Change the Size of an Element
Applied Visual Design: Use the CSS Transform scale Property to Change the Size of an Element
To change the scale of an element, CSS has the transform property, along with its scale() function. The following code example doubles the size of all the paragraph elements on the page:

p {
  transform:scale(2);
}

Increase the size of the element with the id of ball2 to 1.5 times its original size.