Make Screen Reader Navigation Easier with the footer Landmark
Applied Accessibility: Make Screen Reader Navigation Easier with the footer Landmark
Similar to header and nav, the footer element has a built-in landmark feature that allows assistive devices to quickly navigate to it. It's primarily used to contain copyright information or links to related documents that usually sit at the bottom of a page.


Camper Cat's training page is making good progress. Change the div he used to wrap his copyright information at the bottom of the page to a footer element.