Apply the flex-direction Property to Create Rows in the Tweet Embed
CSS Flexbox: Apply the flex-direction Property to Create Rows in the Tweet Embed
The header and footer in the tweet embed example have child items that could be arranged as rows using the flex-direction property. This tells CSS to align the children horizontally.


Add the CSS property flex-direction to both the header and footer and set the value to row.